rename SlideIframe
