+++ src/components/Main/SlideIframe/SlideIframe.css
@@ -1,4 +1,4 @@
-.desktop-slide-iframe__container {
+.slide-iframe__container {
   position: relative;
   width: 1000px;
   height: 1000px;
@@ -6,7 +6,7 @@
   justify-content: center;
 }
 
-.desktop-slide-iframe {
+.slide-iframe {
   width: 100%;
   height: 100%;
   border: none;
@@ -15,13 +15,13 @@
 }
 
 @media (max-width: 1100px) {
-  .desktop-slide-iframe__container {
+  .slide-iframe__container {
     width: calc(100vw - 50px);
   }
 }
 
 @media (max-width: 480x) {
-  .desktop-slide-iframe__container {
+  .slide-iframe__container {
     width: calc(100vw - 16px);
     height: max-content;
   }
