+++ src/components/index.js
@@ -62,7 +62,7 @@ export { RichTextEditor } from './RichTextEditor';
 export { default as MainNavigationBar } from './MainNavigationBar/MainNavigationBar';
 export { default as MainSlider } from './Main/MainSlider/MainSlider';
 export { default as DesktopSlide } from './Main/DesktopSlide/DesktopSlide';
-export { default as DesktopSlideIframe } from './Main/DesktopSlideIframe/DesktopSlideIframe';
+export { default as SlideIframe } from './Main/SlideIframe/SlideIframe';
 export { default as ArticlesTree } from './ArticlesTree/ArticlesTree';
 export { default as TreeList } from './ArticlesTree/TreeList/TreeList';
 export { default as TreeItem } from './ArticlesTree/TreeItem/TreeItem';
